Time: 46 ms (52.86%), Space: 18 MB (65.94%) - LeetHub

+++ 0268-missing-number/0268-missing-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int missingNumber(vector<int>& nums) {
         int n = nums.size();
-        long long total = n * (n+1) / 2;
+        int total = n * (n+1) / 2;
         for(int i: nums){
             total -= i;
         }
